backend/pyproject.toml
[project]
name = "rag-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agenticwrapper",
    "aiosqlite>=0.21",
    "chromadb>=1.0.7",
    "fastapi>=0.115",
    "httpx>=0.28",
    "litellm==1.67.6",
    "loguru>=0.7.3",
    # "magic-pdf[full]==1.3.6",
    "modelscope>=1.25.0",
    "ollama>=0.4",
    "onnxruntime>=1.22.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2",
    "pymupdf>=1.25",
    "pymupdf4llm>=0.0.21",
    "pytest>=8.3.5",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "rapidocr>=2.1.0",
    "requests>=2.32.3",
    "surya-ocr>=0.14.2",
    "uvicorn>=0.34.2",
    "wikipedia>=1.4.0",
]

[tool.uv.sources]
agenticwrapper = { git = "https://github.com/lxl66566/AgenticWrapper" }
